DiscordBotControl/requirements.txt
discord.py>=2.3.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
loguru>=0.7.0
psutil>=5.9.0
pytz>=2023.3
colorama>=0.4.6
Pillow>=10.0.0
PyYAML>=6.0
asyncpg>=0.28.0
aiosqlite>=0.19.0
aiohttp>=3.8.0
aiosqlite>=0.19.0
asyncpg>=0.28.0
colorama>=0.4.6
discord.py>=2.3.0
loguru>=0.7.0
Pillow>=10.0.0
psutil>=5.9.0
python-dotenv>=1.0.0
pytz>=2023.3
PyYAML>=6.0